Refactor: Interact base class handles interact logic
OnInteract method is no longer abstract.
It checks for the IsInteractable and isPlayer flags to determine if it should invoke the InteractHandler event. It also checks if the OneTimeUse flag is true, if so then all IInteractEffects events unsubscribe from the InteractHandler.

Added OneTimeUse flag to check if the InteractHandler be invoked once.

Unregister method added to unsubscibe from the InteractHandler. Called on OnDisable to avoid exceptions.